The ability to configure initial bundle start levels.
Initial Bundle Start Levels are configured through a custom bean in the
Deployment Descriptor, which looks like this:
  <start-level xmlns="urn:jboss:startlevel:1.0">
    <bundle symbolic-name="abc" version="1.2.3" start-level="5"/>
    <bundle symbolic-name="def" start-level="4"/>
  </start-level>
Not specifying a version for a bundle applies the start level to all
versions of the given bundle.

Also moved the Start Level Service out of the
jboss-osgi-bootstrap-system.xml file and into jboss-osgi-bootstrap.xml,
effectively moving it into user space.
The Start Level Service is an optional service, it is enabled by providing
the <start-level> tag in jboss-osgi-bootstrap.xml.

Additionally, the order in which the bootstrap deployment descriptors
are processed is now reversed into:
  META-INF/jboss-osgi-bootstrap-system.xml
  META-INF/jboss-osgi-bootstrap-container.xml
  META-INF/jboss-osgi-bootstrap.xml
  jboss-osgi-bootstrap.xml